Registerを消してLoginに統合
